+++ rust/src/trade/requests/submit_order.rs
@@ -10,7 +10,8 @@ pub struct SubmitOrderOptions {
     symbol: String,
     order_type: OrderType,
     side: OrderSide,
-    submitted_quantity: Decimal,
+    #[serde(with = "serde_utils::int64_str")]
+    submitted_quantity: i64,
     time_in_force: TimeInForceType,
     #[serde(skip_serializing_if = "Option::is_none")]
     submitted_price: Option<Decimal>,
@@ -37,7 +38,7 @@ impl SubmitOrderOptions {
         symbol: impl Into<String>,
         order_type: OrderType,
         side: OrderSide,
-        submitted_quantity: Decimal,
+        submitted_quantity: i64,
         time_in_force: TimeInForceType,
     ) -> Self {
         Self {
